feat: use ApiClientTrait
