+++ src/Controller/Component/ParentsComponent.php
@@ -14,8 +14,7 @@ declare(strict_types=1);
 
 namespace App\Controller\Component;
 
-use BEdita\SDK\BEditaClient;
-use BEdita\WebTools\ApiClientProvider;
+use App\Utility\ApiClientTrait;
 use Cake\Controller\Component;
 
 /**
@@ -24,26 +23,7 @@ use Cake\Controller\Component;
  */
 class ParentsComponent extends Component
 {
-    /**
-     * BEdita Api client
-     *
-     * @var \BEdita\SDK\BEditaClient
-     */
-    protected $apiClient = null;
-
-    /**
-     * Get API client.
-     *
-     * @return \BEdita\SDK\BEditaClient
-     */
-    public function getClient(): BEditaClient
-    {
-        if ($this->apiClient === null) {
-            $this->apiClient = ApiClientProvider::getApiClient();
-        }
-
-        return $this->apiClient;
-    }
+    use ApiClientTrait;
 
     /**
      * Add parent to a folder.
